requirements.txt
alembic==0.9.8
astroid==1.6.2
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.9
docopt==0.6.2
Flask==0.12.2
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
httpserver==1.1.0
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
psycopg2==2.7.4
pycryptodome==3.5.1
pylint==1.8.3
python-dateutil==2.7.0
python-editor==1.0.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.5
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
